organize code for better logic
1. Move runbook related init code into runbook.py
2. cleanup commands.py, make it like a command wrapper.
3. move test case selector to lisarunner to centralize flow.
4. other minor improvements.
